Implement synchro reconfiguration

+++ cream-administration/cream-administration-api/src/main/java/fr/liglab/adele/cream/administration/api/AdministrationService.java
@@ -12,7 +12,7 @@ public interface AdministrationService {
 
     ImmutableContextEntity getContextEntity(String id);
 
-    void reconfigureContextEntityFrequency(String contextEntityId, String contextStateId, int frequency, TimeUnit unit);
+    void reconfigureContextEntityFrequency(String contextEntityId, String contextStateId, long frequency, TimeUnit unit);
 
     void reconfigureContextEntityComposition(String contextEntityId,String functionnalExtensionId,String functionnalExtensionImplementation);
 }
